Prototype of repology-vulnupdater reimplementation
Compared to python implementation, this does not update vulnerable_cpes
table and doesn't handle CPE dict, however it produces equal cves table
content.

Let this commit serve historical purposes, as it turned out that NVD is
going to discontinue CVE feeds in favor of more flexible API, so we'll
have to rewrite the update process based on existing code.

The code added here and will be removed, which may be of particular
interest, is at src/update.rs:124 and it demonstrates combination of
async stream handlers (e.g. convert reqwest stream into async reader
and handle stream decompression on the fly).