webapp: Hide tokio unstable metrics under #cfg
Hide usage of unstable tokio runtime metrics under the same cfg tokio uses.
This fixes build when rust flags are set (in which case cargo won't append
`--cfg tokio_unstable` to rustflags and metrics won't be available). In
particuler, it fixes CI because it is build with flags set to
-Cinstrument-coverage.