ci: Fix caching
Fallback to the setup from actions's example.

It turns out that on cache hit cache is not stored in the end of a job,
so dedicated cargo cache becomes useless for it's not filled with new
and updated crates. Thus we always need hashFiles in the key.

With restore-keys we could avoid cache invalidation on Cargo.lock
change, but it would instead make the cache to grow infinitely as
nothing is removed, which is not what we need since it slows things
down.

So instead, allow the cache work only between Cargo.lock changes, which
speeds up consequent builds and doesn't produce much overhead.