webapp: Fix handling out of range segment on graphs
Before this, graphs didn't include the segment which intersects left
boundary of a graph (21 day boundary in other words). This was not
visible because on 21 day range, updates are rather frequent, and the
leftmost segment starts just after the boundary.

However, if there's huge gap in updates, or history is deduplicated
to remove points which do not have any counter changes (see #120),
we'll see a missing segment. Fix that by fetching an extra data point
just before the graph range.