webapp: graphs: ignore null values properly
Filter out NULL values _after_ requesting them. This affects
out of range subquery (one with LIMIT 1), which, having filter
condition, may, instead of returning single row, start scanning
the whole table. This is not desirable and is unlikely to return
any data, as metrics tend to be added rather than remived, and
if metric didn't exist in moment T it's unlikely to exist before T.