webapp: Implement basic snapshot tests (#75)
This was supposed to work on production data extract, but it turned out
that there's no easy way to extract data conditionally and put it in for
of sql insert statements. So instead, just take fixtures and urls from
our integration tests and run snapshot testing on these.

IMO it's too early to put snapshots into the repository, so use case for
these tests is mostly local development, and for this purpose they are
hidden between snapshot-tests feature, and snapshots directory is added
go .gitignore. To use them, just run `make snapshot-tests` before and
after your changes.

While here, switch integration test fixtures to fixed timestamps to make
them suitable for snapshots.