webapp: Wrap state in Arc
This reduces amount of work needed for request handling, which involves
cloning a state, as we now only do a reference count increment and do
not touch internals of AppState (which may have multiple Arc's, or
heavier to clone data inside). This also allows to store non-clonable
stuff inside AppState, such as Mutexes for mutable state. We also lo
longer need to wrap FontMeasurer in Arc.

This is a prerequisite for making RepositoryDataCache API synchronous
(#49).