webapp: Fix appearance of related project links on project/related page
Switching to displaying related project links (∗) introduced a minor
regression. We can't display the link unconditionally, because it is
possible to view related projects page for a project which has no
related projects, in which case there would be only one project on
that page and it doesn't need the related link.

Issue: #124