webapp: Make (part of) tests more granular
This simplifies handling failures, as it's apparent which exactly test
have failed. This is especially useful for snapshot tests, for which all
snapshots may be invalidated at once, and accepting new ones would
require N test-accept loops where N is the number of test cases in a
test.

On the downside, this produces a lot of boilerplate, and slows down the
test, as heavy database migrations need to be applied N times more now.