ci: Another attempt to set up cache action
Dependency crates are rebuilt each action run, so apparantly cache
doesn't work well. The likely cause is that we build on rust-nighly
which changes almost every day, but cache is not aware of that, so
unless Cargo.lock also changes, a cache hit would be counted, but
since rust version changes, it would be rebuild from scratch. So:

- Split cache to cargo and target.
- Cargo cache contains downloaded crates and is not dependent on rust version.
- Target cache depends on rust version.
- Also add restore-keys to reuse partially matching caches.

How this should work:
- Target cache resets on each rust version bump, which also prevents it
  from uncontrollable growth, so it's safe to use restore-keys eagerly.
- Cargo cache is not affected by the rust version. Since it's also
  restored, we should observe whether it's prone to growth, but it
  shouldn't be too big since dependencies do not change often.