webapp: Implement maintainers endpoint (fixes #159)
This implementation simplifies pagination: we no longer query project
names range and disable pagination buttons when we hit start or end
of the list. This imposes unwanted overhead of fetching a lot of extra
rows (especially when search feature is used), and it doesn't seem to
have worked in python impelementation anyway.