webapp: Update index page important projects asynchronously
These are the heaviest projects, and fetching/processing them take a
lot of resources, we cannot afford to do it in the index page handler.
So instead, implement a background task to update these each 10 minutes.

The code is not completely clean, for instance we're using fn and struct
right from the view code, while these should've been moved into separate
module, but that's probably OK as we're going to rewrite index page and
get rid of that projects listing anyway (#93).

Related to #160.