linkchecker: Implement link checker
Notable improvements/changes over python implementation:
- Politeness delays are considered for each request in a redirect
  chain (not once for a whole chain).
- Only 200 responses are now counted as success (there's marginal
  amount of 201-204 responses which we don't want to handle).
- Redirects are now only saved for 200 responses.
- Check duration is no longer saved to the database: new link checker
  exposes enough statistics through logs and metrics, and wasting
  database space on not used data does not make sense.
- Fetching tasks from the database is now much more effecient and
  tunable.
- When a host queue is overflown, we now postpone excess check tasks
  not to needlessly iterate over them again and again.

The most fun part of this implementation is that HTTP client is
implemented in Python (in fact, it's part of the old python link
checker code) because of it's better error introspection capabilities,
and also to explore possibilities of async communiation with async
python process which we may need later in the updater rewrite. Here
it will be replaced with reqwest eventually.

Fixes #249