linkchecker: Fix Feeder loop restart condition
It's very likely that between batch fetches, new links will become
eligible for fetching, so every batch will return some tasks, even
if the cursor is very close to now moment, so the feeder will never
restart the loop and visit urls with older next_check times which
may appear in the table due to various reasons. So restart the loop
as soon as we can't fetch the full batch from the database.

Fixes #256