linkchecker: Implement requeue cooldown
There's a rare case when an url may be queued for checking twice:
- Url is being processed
- New batch is fetched, which contains the same url
- Url completes processing and is thus removed from bucket queue and task_ids
- Url from the batch is queued again

Generally this is rare and harmless case, however now it affects failure
streak counter (#279) by counting two failures in a row. This makes
failure streak less reliable, as rechecks are expected to be separated
by sufficient time interval.

To avoid this, introduce a mechanism similar to already existing task_ids
hashset which prevents douple queuing. Recently processed ids are stored
in a hashset, ands ids present in that set may not be queued again. The
set is cleared manually before each batch. As a result, tasks which
completed during queuing of a consequent batch, may not be queued from
that very batch.