linkchecker: hosts: Bump timeout for git.pld-linux.org
It calls git multiple times per request and may respond in up to 10
secs. Waiting for it instead of timing out earlier makes more sense
as long as fast recheck is enabled (it currently is) - we either wait
up to 30 secs until success once, or we wait up to 10 secs for each
failed retry, which sums up into more time.