linkchecker: disable HTTP retries
reqwest 0.1.23 now allows to customize retry policy. The default has
not changed and is to only retry "low level protocol NACKs", which,
as explained in [1], implies recreating presistent HTTP/2 connections
transparently when needed. That mostly does not affect us as we create
a new Client for each request, and thus do not have any persistent
connections, but still
- be explicit that we don't expect any retries.
- handle the rare cases that the request gets such low level protocol
  NACK right away - report an error, do not retry

[1] https://github.com/seanmonstar/reqwest/pull/2763#issuecomment-3200851949