updater: Make parser accept a function
It looks more covenient for parser to accept an arbitrary processing
function (in form of &mut dyn FnMut) to pass parsed package data to
instead of Sink trait, which implies that a dedicate type needs to be
written for each processing case, with questionable composability.

Get rid of all Sinks in the process.